refactor: capitalize spark/src structure
- Partially addresses #40
