+++ libs/spark/src/form-inputs/index.ts
@@ -1,8 +1,8 @@
-export { Menu } from './menu';
-export { Checkbox } from './checkbox';
-export { Input } from './input';
-export { Radio } from './radio';
-export type { MenuProps, MenuItemOptions } from './menu';
-export type { CheckboxProps } from './checkbox';
-export type { InputProps } from './input';
-export type { RadioProps } from './radio';
+export { Menu } from './Menu';
+export { Checkbox } from './Checkbox';
+export { Input } from './Input';
+export { Radio } from './Radio';
+export type { MenuProps, MenuItemOptions } from './Menu';
+export type { CheckboxProps } from './Checkbox';
+export type { InputProps } from './Input';
+export type { RadioProps } from './Radio';
