+++ libs/spark/src/icons/icons.tsx
@@ -1,5 +1,5 @@
 import React, { FC } from 'react';
-import { Icon } from './icon';
+import { Icon } from './Icon';
 import {
   CaretUp,
   CaretDown,
