+++ libs/spark/src/illustrations/Illustration.tsx
