+++ libs/spark/src/logos/Logo.tsx
