+++ libs/spark/src/typography/index.ts
@@ -1,2 +1,2 @@
-export { default as Typography } from './typography';
-export type { TypographyProps } from './typography';
+export { default as Typography } from './Typography';
+export type { TypographyProps } from './Typography';
